+++ src/pkgIndex.tcl
@@ -1,4 +1,4 @@
-package ifneeded alited 1.8.13 [list source [file join $dir alited.tcl]]
+package ifneeded alited 1.8.14 [list source [file join $dir alited.tcl]]
 
 namespace eval ::alited {